Refactor SCSS to allow toggling depressed styles with a class
Svelte strips "unused" classes from compiled CSS.  The accepted workaround for this is to use `:global` to eject the styles, but (optionally) use qualified selectors to keep the local scoping.  In this case it's more awkward still due to the involvement of SCSS and pseudo classes.

This could be avoided by just hard-coding the markup and then the `depressed` class could be toggled using the more idiomatically svelte-y `class:depressed={condition}` on each key <span/>.  This may be a better solution in the final analysis.