Track when panning viewport
Without this, zooming out while playing causes the x coordinate of the viewport to get out of sync and the roll image to fly off the screen... :)

The other solution I found was to set `animationSpeed` to `0`, which would be a good solution imo if it could be applied only to zooming and not to panning, but they appear to both use the same setting.  Disabling the animations altogether seems a shame.