Overlay to make white parts of roll images match the background
The white parts don't look great to me, and this (or something very like it) is more-or-less the only solution available to us to remove them.  Depending on the colour used, it does affect the colour of the roll itself too (the darker the background color, the greater the effect), but for my zwei pfennig I reckon it's worth it.