Add holeData to JSON files (where available)
These are immediately very heavy now, even once filtered for the keys (I think!) we actually need, and with the *px strings converted to integers.  Not pretty-printing them for production saves a little bit (~10%?), and they'll compress well over the wire, of course, but still it might be worth being terser.