Weed out "un-highlight-able" holes before runtime
Holes with no NOTE_ATTACK and OFF_TIME are the "continuation" holes where a key has been held down.  The duration is encoded in the NOTE_ATTACK and OFF_TIME values of the first hole in the sequence.  Data for the subsequent holes do include a WIDTH_ROW which, I assume, is a vertical dimension and could perhaps be used for something, but it's not needed for the present implementation, and dropping such unused data reduces the size of the JSON files that need to be fetched by ~60% across the development corpus.

See also https://github.com/simonwiles/pianolatron-data/commit/cfc952d86e28af3b53ffcda371d2e627ce30bbd0.