Set default <mark/> styling
`<mark/>` elements pick up a background colour from the browser stylesheet, and so orphaned `<mark/>`s (see previous commit) result in a visual highlight.  This hides the problem, rather than fixing it, tbh -- but I think there's always going to be a bit of flakiness around the `mouseout` events, and this (and the previous commit) is  tidy enough to be acceptable, imo.