Refactor to a separate <RollViewerControls/> component
The motivation behind this (aside from neater code in general) is that I would like to use event listeners to handle when the buttons should be disabled, and I'm worried that the event listeners might be expensive.  With this refactor the component will be unmounted when not needed, and the event listeners can be destroyed.