Better styling of activeNotesDetails
Yikes this is getting to be some rather gnarly (S)CSS.

This commit shifts the `mix-blend-mode: multiply`'d overlay to a `::before` pseudo-element to avoid having to `mix-blend-mode: multiply` the whole element (which can't then be "wound off" on the `::after` pseudo-element).

There's some funky behaviour around applying the blend mode whereby it doesn't seem to work as I think it should when the hover styles are applied.  It looks kinda buggy to me, tbh, but gecko and blink both do the same thing, so perhaps there's something in the spec. I'm not understanding.  In any event, effectively destroying the `::before` pseudo-element on `:hover` creates the same effect as previously (the yellow "highlight" is removed when the `darkturquoise` outline is applied), which I think is acceptable and arguably even good.  Note too that the choice of using `position: relative` to do it (rather than, say, `content: none` which might be semantically preferable) is a deliberate one, and connected with the same odd and possibly buggy interaction of `mix-blend-mode` and `@keyframes`.