[WIP] show info from MIDI file on hover overlay
There is a problem here; there are discrepancies between the `TRACKER_HOLE + WELTE_MIDI_START` values that are calculated and the `noteNumber` that's returned by midiplayer-js.  This came to light when trying to determine which midi note event corresponds to a given hole -- when there's only one event that coincides with the tick 'location' of the hole, this is assumed to be correct.  When there's more than one, however, the `TRACKER_HOLE + WELTE_MIDI_START === noteEvent.noteNumber` comparison is the only (?) way to determine which of the events corresponds to a given hole.  Except it doesn't always match up.

This commit spits out examples to the console.  The number of examples varies greatly by roll.