Fix SCSS imports (again)
This feels like a bug (especially when it works just fine with `snowpack dev`), and, frankly, there seem to be a few of those w/r/t snowpack and `@snowpack/plugin-sass`, tbh, but I'm too tired to get into it now, and this addresses the immediate problem satisfactorily.  (it's 100% safe because it's build-time compilation, so if the assets are built correctly, it works -- no sneaky edge cases here!)

Worse, there are additional issues related to the SASS/SCSS setup related to upgrading both Snowpack *and* Svelte :(  But these don't need to be tackled immediately, at least.