Except reactive expressions from linter rules
I think the idiomatic Svelte-yness of these expressions justifies the exception.  Note that the `eslint-disable` syntax used here disables the rules until the end of the script block (the end of the file, in this case).  Since these reactive statements come at the end of the file anyway, this is neat and tidy.