Change spelling of pedalling to pedaling
There seems to be a vague suggestion that two *l*s is a British spelling and one is an American spelling, but both Merriam-Webster and the OED list both as variants with no mention of a transatlantic significance.  My instinct is to use two, but I noticed the prototype has one, so perhaps it *is* a po-tay-to po-tah-to thing? :shrug: