Refactor playing state control
The impetus for this refactor was a small circular dependency between the (old) `playPauseApp()` function and the `skipToTick()` function.  It wasn't a causing an immediate problem, but it was a bit of a code smell, and an indication that the player API wasn't as clean as it could be.  This is much better now.