Rework SCSS  for dynamically created element and styles
Something about the Svelte upgrade from 3.29 to 3.38 broke a load of my `:global` styling.  Tbh I don't really understand why (the way this all works is my least favourite thing about Svelte by a distance), but it's one of those things where if the CSS it spits out at the end of the day is correct, then it's working, so it doesn't unnerve me too much that I don't really understand *why* this change works (or is necessary in the first place).  It does work, so it'll do.