preventDefault when closing the dropdown
This is intended to yield better behaviour when clicking on the roll while the dropdown is open:  previously nothing would happen as events on OSD call `preventDefault()` (I think?) -- this way around the dropdown will close.  To prevent the click from generating a zoom event, I've simply disable all `clickToZoom` in OpenSeadragon (alternative solutions are available, but orders of magnitude more involved, and I personally find the `clickToZoom` more annoying than anything else, so I thought I'd test the waters... :)).

Concern:  I'm not sure why `defaultPrevented` was being checked previously, tbh (:S) -- it's possible this was for a good reason...