Set fixed styles only once
OSD writes these directly to the element, so they kinda need to be altered there.  I did consider putting these styles in the (s)css, but it would have meant either copious use of `!important` or setting the values to `unset` in the JS anyway, so this is what I'm going with.