Handle mousewheel zooming manually
OSD's mousewheel zooming goes crazy if used while the app is playing (I think it's trying to be clever with it's mixed panning or zooming -- in any event, it's really unhelpful in our case).  After trying a number of other solutions, seems that just disabling it and handling the scroll events ourselves produces the best outcome.