Prevent shifting playback position to the middle on loading
This is a hack to prevent the `<RollViewer/>` from causing the playback position to advance immediately upon opening.

OSD opens in the image in the middle.  The kludgy solution offered by the main OSD author (e.g. [here](https://stackoverflow.com/a/30264133/88411) and [here](https://stackoverflow.com/a/63060953/88411) and [here](https://github.com/openseadragon/openseadragon/issues/979) is no good to us because two (yes, two) `pan` events are fired even before the `open` event :(

Fortunately for our case, those two pan events are called with `immediately=true`, and since we're never doing that ourselves anymore, a hacky solution (for now at least) is just two use that argument as a proxy to identify those two spurious events and ignore them.