Use OSD's `scrollToZoom`
OSD can exhibit odd behaviour if the nav strip is used to jump around while zoom animations are playing (see https://github.com/sul-cidr/pianolatron/pull/68#issuecomment-896304339).  One way to solve this is to use OSD's own `scrollToZoom`, which doesn't operate when scrolling on the nav strip.

This is effectively a reversion of 5b10d553.