Allow panning beyond image bounds when dragging on the nav strip
OSD prevents panning beyond the top and bottom of the image when using the navigator -- but we need to be able to do so so that the tracker bar position (i.e. the play position) can be reach the extremities of the playable range (i.e. the holes) when the viewport is zoomed out.