Add support for (rudimentary) tooltips
I'm not sure to what extent a) we'll want to use this; and b) an implementation this simple will cut it.  On the one hand I'd prefer something CSS-only (so that it can easily be used on no-js pages); on the other, we might need more positioning logic than this allows.