Persist other `$userSettings` in localStorage
`theme` is a little more complicated because it needs to be set somehow when the state is loaded (if it exists in localStorage in the first place, of course) -- it's only reactive if the `<AdvancedSettings/>` component is in the tree.

Since this ideally needs to be done before the app loads anyway (to prevent the default theme snapping to the custom theme in an unseemly fashion when the app code is parsed), I've moved things around a bit.  Applying the `data-theme` attribute to the `documentElement` in `base.astro` is nice because it means we don't even have to wait for `document.body`.  It will also allow preserving the theme setting for static pages, if that's deemed desirable.  Moving the relevant CSS to `themes.scss` is just for neatness.