Use controlConfig object to set slider attribs
This is good for consistency and maintainability (as evidenced by the fact that I found an *inconsistency* while doing this), but one user-facing change in the immediate instance is that the tempo slider is less fine-grained than before.  I'm inclined to think this isn't a problem at this point, but it may bear some discussion.