Prevent inconsistent editing state
Only allow one shortcut to be in editing state at one time; reset all to `editing = false` when "reset all" is invoked.  Ensure all are set to `false` when the dialogue is closed (`shortcut.editing` is persisted in `localStorage`, and we don't want the dialogue opening back up in a "editing" state).