Preload the CIDR logo for the welcome screen
I have slightly mixed feelings about this.  One the one hand, the http request for the image gets stuck behind other requests, so the image can show up super late, hence the fading in and so on.  On the other, this is not much of an issue on fast connections with http2, and using the `preload` link means the image is loaded on every page load, even if the welcome screen isn't shown.  Back on the first hand again, though, it's almost inevitable that it will be cached.  So on balance I think I'd rather have this in :)