"Reset All Settings" needs some UI feedback
So it's time to add a little more functionality to the Notification component.  This is not going to see a tonne of use, tbh, so I'm loathe to spend a tonne of time polishing it (yes, both use and time are measured in tonnes), but hopefully this doesn't have too many rough edges.