Generic fix for the `<RollViewer/>` not always updating
This is bit of a kludge to address the problem of the `<RollViewer/>` component getting into an inconsistent state if it's not destroyed properly when a new roll is selected.  The problem is caused when `<RollViewer/>`s lifecycle `onDestroy` method is not called, which can happen when it's removed while one or more of its children are in the middle of a `transition` (see https://github.com/sveltejs/svelte/issues/5268).

One option would be to remove all Svelte `transition`s on child elements (the "Downloading roll image..." message, `<RollViewerControls/>`, and `<RollViewerScaleBar/>`) but this seems a shame.  In my testing, this fix seems to address the problem without needing to remove the `transition`s.