Recreate `<KeyboardShortcutEditorRow/>` reset button when `isChanged` is updated
Another (perhaps better, certainly more generic) solution would be to have the action return an `update` method that takes care of dynamic changes to the `text` prop, but this is rather more complicated than it might at first appear (you need to bind functions that close over the node and the component instance, and because the input props are object attributes the update events need filtering and &c, and &c).

This solution is neat and meets the present need.