Use already existing holesByTickInterval
This commit calculates the tick range for a given SVG partition  (a bit of a head-bender!) and then uses the `holesByTickInterval` interval tree to pass only those holes to the `createHolesOverlaySvg` function.

It also uses the beginning and end of the hole-containing range for the overlays, rather than the beginning and end of the image.

The hole thing is made complicated by the way that `FIRST_HOLE` and `LAST_HOLE` correspond differently to image coordinates depending on `$scrollDownwards`, and I've been resisting the temptation to rethink/regularize this a bit.  I think probably leave it alone is better, but...