Remove transitions from right sidebar panels
For the sake of getting this branch commiited, I've just removed the (now inadequate) slide-in/out animations.  I'm doing a cleanup of this tabbing arrangement, and animations may get added back in (or they may very well not -- animations of this kind of UI aren't common for good reason.  Might try a simple cross-fade or something.).