Use a new typeface for all text
This was initially motivated simply by the fact that FF and Chrome use different size typefaces for `serif` and `sans-serif` on my system, and it was making the layout quite different on each.  This typeface seems okay, but the higher priority was to have the *same* typeface on as many systems as possible.