Separate out `$pianoSettings` store
Reverts some of the changes in d8c75937.  The stores were grouped partly just for convenience, but also partly because `createPersistedStore` was written in such a way that it would only handle objects, so this commit also includes an update to that function so that it can handle simple values too.