Increase the size of the spliner "control points"
I find the points extremely fiddly to get hold of (especially the ones at the edges), and so even though it's a little unsightly I actually prefer this value set to `10` or even higher.  However, I notice you explicitly lowered the value to `5` from its default of `8`, so I propose a compromise of `7`, chosen because it makes the circles look more-or-less the same size as the ones on the sliders.