Prevent brief flash of "undefined" on error message resolution
To keep the message there and fading out (rather than just snapping it out of existence by rendering `{errorMessage || ""}` or something like that) this uses a Svelte transition.  The mix of fading-in/-out implementations is a little bit of a dog's dinner at this point, but each technique has its pros and cons, and each has its use-case I suppose.