Update SASS syntax
`@import` is deprecated as of 2021/10 and slated for removal in 2022/10, so this gets out ahead of that.  This commit probably obviates the need for 094bf381, if I understand correctly, but that was a small (if now unnecessary) improvement anyway.

This isn't really the way the new SASS modules architecture is supposed to work, but it's the lowest-touch response to the deprecation, and is plenty acceptable for this project imo.