Use strings corresponding to the user's keyboard layout where possible
I had originally hoped this might give us `Win`/`Cmd` instead of `Meta`, or something like that, but it doesn't :(

However (and despite a lack of broad support) this is probably worth having for two reasons: 1) it will yield the correct string on e.g. a QWERTZ keyboard or a Dvorak keyboard; and it ignores the effect of <kbd>Shift</kbd>.  w/r/t the latter, prior to this commit (and on browsers where this API isn't available) trying to assign a command to <kbd>Shift</kbd> + <kbd>3</kbd> results in `#` being used as the label (on a US keyboard -- presumably `£` on a UK keyboard) but the command is actually bound to the <kbd>3</kbd> key, with or without <kbd>Shift</kbd>.

Behaviour on browsers without support for this API is unaltered.