Handle `$activeNotes` in `startNote` and `stopNote` functions
It shouldn't be the job of the `<WebMidi/>` component to manage this.

However, this does affect the way the `<Keyboard/>` component (the other client for these functions) works a little -- clicking the keys now removes the `depressed` class if it was previously applied (i.e. if the roll player state is in the middle of (a) note(s)).  This is in line with the behaviour when using the keys on a WebMIDI-connected device.

If this is considered undesirable (and I think I'd lean that way, if not perhaps all that hard), then both the on-screen keyboard and WebMIDI-connected devices could could be decoupled from `$activeNotes`, but I'd rather do that with a argument to the `startNote()` and `stopNote()` functions (similar to the `fromMidi` argument).