Refactoring
This is some opinionated stylistic refactoring.  The main changes introduce more destructuring and switch statements (which are technically faster, but the point here is more about readability).  I think I can make the case that all of these changes are in line with modern best practices, but I'm happy to entertain counter arguments :)