Update catalog metadata format and regex for label selector display
Licensee and Duo-Art rolls have roll numbers that aren't just digits (e.g.,
B-6294) and multiword publisher names (De Luxe) that don't play well with the
existing code to parse and format the entries for the rolls in the roll
selector. The solution here is to explicitly label the roll numbers and
publishers in catalog.json and then streamline the formatting regex accordingly.