Realism enhancements: tempo/timing and tracker width emulation
Many of the edits are just explanatory comments for the red Welte expression box
code. The actual functional changes involve emulating the "prolonged suction" of
the tracker bar holes as they move over the end of each perforation, and
accurately modeling crescendo/decrescendo rates when the tempo is updated in the
middle of such an event (due to acceleration emulation).