Support for 88-note in-app expression & viz
88-note rolls really only have one expression control (the "snakebite" accent
holes), plus the sustain pedal, but the operation of the former is considerably
different from most of the red Welte controls, and implementing the expression
box for the 88-note rolls turned out to be more complicated than expected. It
also revealed some additional needed improvements to the red Welte expression
box code. Hopefully having implementations of both will help to highlight the
portions of the expression functionality that should be abstracted out and
shared between different expression boxes.