Implement tempoMap as an IntervalTree for all expression boxes
The tempo lookup operations on the previous linear tempo map were causing noticeable playback hiccups, especially on longer rolls when using the in-app expression boxes, which implement more fine-grained acceleration emulation (via frequent tempo updates) than the tempo updates that are baked into the expressive MIDI files. This change also necessitated modifications to how ticks are converted to milliseconds (and vice versa) when calculating the expression curves.

Note that further work is needed to ensure that the playback tempo immediately changes to the proper setting when jumping to a different playback point on the roll -- this is needed for all expression boxes, including the expressive MIDI. The 88-note and welte-red expression boxes are likewise missing this functionality for pedal events, although this is provided for expressive MIDI.

This update also applies the streamlined trackerbar extension emulation from the previous commit to 88-note rolls.