Performance improvements when calculating expression visualizations
This method of finding the minimum and maximum note velocities for the purpose
of establishing the hole color range is several orders of magnitude faster than
loading all of the values into a set, which was causing multi-second delays on
long rolls (e.g., hm523dq5554) and especially with in-app computed velocities,
which can (and arguably should) have arbitrary precision, compared to the
integer-only values available from expression MIDI files.

Drawing the expression curves, even on long rolls, turned out not to be the
culprit for these slowdowns, and does not seem to need any further optimization
(such as splitting them into segments and only displaying the ones that are in
the viewer window, as is done for the holes).