Add expression menu with in-app toggle and expression parameters display
For all known roll types, we'll just want to be able to choose between using the expression MIDI file (when available) and the in-app expression logic for that roll type (also if available), so the new menu provides that choice within the framework of the multiple expression boxes, although its implementation could be cleaned up a bit.

The expression parameter display is again available via this menu, when in-app expression is selected. More work is needed to make the playback and visualizations actually reactive to changes in the parameters, however. For now, only the "Reset to Defaults" button is even minimally functional.