Recompute derived exp params when tunable params are changed
This change is crucial to achieving full in-app expression emulation, whereby changing any "tunable" parameter via the Expression Settings menu automatically triggeres an update of the non-tunable variables derived from them, which are then used to recompute the expression curves, assign new velocities to notes, update overlays, etc.

Previously, the "derived" values were exposed and locked to the numbers displayed in the Expression Settings menu, which wouldn't automatically update when the "tunable" parameters were changed -- so the visualizations and velocities would only respond to changes in a limited, superficial way.

The process of recomputing the values and updating the velocities and visualizations is noticeably slower now that the functionality is correct, but it's not clear that this can be made to run much faster.