Use Web Audio API to record the ToneJS piano
The recording controls are stuck in under the MIDI tab, which may need to change eventually. Their styling is pretty rudimentary, and the way the recording controls and functions are wired together also probably could be done more elegantly. But this at least demonstrates all of the basic functionality one would desire of a recording feature: the user can start, pause, clear, and export a recording to a file, and the UI provides some basic information about the progress of the recording, including the duration.