Simplify controls and recording options
Recordings made with the MediaRecorder interface of the MediaStream Recording API that include multiple stops and starts (or pauses and resumes, using the appropriate MediaRecorder methods) tend not to play properly when exported as encoded sound files. So for now, it's preferable for the in-app audio recording controls only to allow single-shot recordings -- so once a recording has been stopped, it can be exported one or more times before being cleared, but it can't be resumed.

It's also likely preferable for all of the recording control buttons to be visible at all times, but disabled/greyed out as appropriate at the various stages of the recording/exporting/clearing process, so that's what is done here.