Draft of potential improvements to the Duo-Art expression box
The most significant change here is that the expanded area of effect of the "theme" snakebite holes is reduced, and rather than being rather messily implemented during the calculation of the expression map, these ranges are added to the control events at the same time and via the same method as the tracker extension emulation -- that is, by adjusting the tick values of the "on" and "off" events for those holes. This change might also be beneficially applied to some of the other expression boxes, such as for 88-note rolls.