Enable specifying alt text for Icons and IconButtons for mouseover
The alt text will appear in default browser tooltips on mouseover. Not supplying an alt= value causes it not to appear, which is the case i.e. for the key bindings menu, when PopperJS tooltips are being used instead. It is unclear whether support for desc text would be worthwhile. Better ARIA integration for the SVG elements remains a possibility.