Linting
`getVelocityAtTime` isn't in the critical path, and the optimization is effectively meaningless for performance, but this may be clearer (since the comment about simplifying it is now removed).

Also, my habit of adding `?` at the beginning of comments comes from using https://marketplace.visualstudio.com/items?itemName=aaron-bond.better-comments, so repeating it keeps this convention.  Not important, of course.