No need to subscribe to changes in expressionParameters
initializeExpressionizer() is now called whenever a significant settings change is made (roll type, expression parameter) so there's no need anymore to listen for changes in expressionParameters and (re)build the note velocities map then.

In fact, it seemed that previous practice was resulting in chains of subscription listeners that caused the note velocities map being built repeatedly in an ever-increasing number of iterations (2, 4, 6, 8...) each time an expression setting was changed. This version of the code no longer has that symptom, and in fact recalculations of the note velocities map on settings changes are now quite snappy.