Improved implementation of expression curve overlays
The original implementation of the expression curves and guides caused severe performance issues for some hw+sw combos because, similar to the original implementation of the hole overlays, it drew thousands of SVG elements on the full length of the roll image that then would need to be animated during playback. This version uses the same partitioning method that was adopted for the hole overlays to break the expression curves into 1000-pixel-tall SVG chunks and then only draw/animate the SVG partitions that are visible in the current viewer window.

The code to do the actual drawing of the expression curve SVGs has been streamlined a bit, and the UX related to the overlays is hopefully improved. The curve overlays draw correctly for both downward (red Welte) and upward panning rolls (the rest), although occasionally the final segments are dropped for some upward panning rolls -- this is most noticeable for 88-note rolls, and will be addressed in time. Duo-Art rolls are not yet supported, because this branch does not yet have fully working in-app expression code for Duo-Art rolls.