Removing duplicated playback buttons and other related changes
The main purpose of this commit was to comment out the extra control buttons for play/pause, record, skip forward/back, etc. because they will be replaced eventually by the larger, improved buttons in the control strip above the roll viewer, which hopefully will be able to take advantage of the improved `<Icon>` and `<IconButton>` components in this PR. The scope of the commit then grew to incorporate tooltips and keyboard shortcuts for the record button(s) and others, before the changes mercifully reached an end.