Semantic table layout for screenreader accessibility
This is an attempt to format a table the "right" way for screenreader a11y; the core notion is that rows as well as columns should have headers (with the appropriate scope/role attributes) so that when a screenreader is reading out the rows, it's easier to grok the primary item to which the entry each row cell belongs. There may be a way to do this less invasively, without reformatting the whole table so that the most ID-like element (in this case, the roll's publisher and issue number) is in the first column as a row header, but if so, it wasn't apparent to me.