Use `--primary-accent` for transport controls hover style
This changes only the icon, leaving the border white on hover, as before.  I still don't love it, tbh, but I think it's unambiguously better than before.
update: added a background too, as without it this change didn't work too well with the grey theme.  I like it more this way, I think 🤷