Move the mode switch in a bit
This is to counteract the impression of it being out of alignment because of the transparent gap on the left of (almost?) all the roll images.  The button now appears inset a little when the roll image is zoomed in (and covering all the space reserved for it), but that's not nearly so niggling as the apparent mis-alignment when the roll loads in "full-width".