Allow game controller to trigger bookmark of roll
The original gamepad button mapping scheme used all of the buttons and triggers on a standard game controller except one, leading to a conspicuously empty slot in the gamepad button mapping graphic in the new help section. So the bookmark feature was mapped to the unallocated button. Being able to copy a roll's permalink to the clipboard from a game controller may not be the most useful functionality, but it's not harmful, and it produces a nicer graphic, so there.