Build in-app expression MIDI using current params
You get much better control over the final output MIDI from the ToneJS MIDI package if you assemble all of the data for it in a big JSON object and then load the JSON into a ToneJS MIDI instance. The actual API for building a MIDI instance on the fly, which is still used in the live-recording code (also in WebMidi.svelte), is much more limited. In this situation, the data comes from the roll metadata and the in-app expression notes, tempo and pedaling interval trees, which are computed at load time and whenever the in-app expression parameters are changed.