Add another piano playback knob: sustain linger
This is a bit experimental, but it helps to address the frequent observation that much of the sustain pedaling on reproducing rolls seems a bit too "dry" (abrupt) even given differing times and tastes, and so it might possibly be due to a shortcoming of the recording equipment. Note also a minor change to ensure that the play/pause icon reverts to the stopped state when playback reaches the end of the notes on the roll.