Roll navigator strip should behave properly after window resize
Just as changes in OSD 4 meant that the progress indicator overlay on the navigator strip wouldn't be drawn in the proper location after the browser window is resized (this was addressed in a previous PR), click and drag events on the strip after a window resize also don't work the way they used to in OSD 3. This change to the handler for both types of events seems to do the trick. Note that this commit probably could/should be a standalone branch and PR.