Update $ticksPerSecond on tempo changes in in-app and MIDI expression
This fixes a bug whereby the "Roll Speed" readout in the ViewerMetrics component won't update on tempo adjustments during playback (which occur due to acceleration emulation) because the $ticksPerSecond store that triggers calculation of the Roll Speed for display is only updated on manual skip/stop/start events. During playback, the tempo setting was only changed in the MIDI player, which is not hooked reactively to the $ticksPerSecond store. A theoretically more elegant approach in which the MIDI player tempo and ticksPerSecond value are all linked reactively to a global $currentTempo store was also tried, but it was oddly not as performant and prone to glitches.