Move pagination controls out of `<tfoot/>`
I wonder what *is* acceptable in `<tfoot/>` element, if not information and controls for what's displayed in the table?  Only stuff that's relevant on a column-by-column basis?  I'll certainly never dare to use `colspan` again.

Also noticed that the `on:keypress` event handlers are superfluous, so removed them in the process.

Addresses SODA #515.