Better header background
The first time I saw this header I wanted to fix this.  Imo this looks tonnes better, but 1) you may disagree? and 2) it isn't exactly the same as the spotlight header (although that ship is already on its way out of port, to some degree).  Either of these considerations may be sufficient to nix this -- it's not very important, it just bugged me a little.