Adjust footer links for smaller viewports
Part 2 of addressing the issue with overflows with very small viewports.  This commit fixes up the `stanfordFooter` component.  As with the sub-header, standard responsive design techniques mean the overflows that were causing the major problem disappear without the need for direct and explicit mediation.  A little more rejigging was required with the footer.

Addresses SODA #516 (though further reponsive work is incoming for the table itself).