Layout for documentation pages should extend base layout
My attention was initially drawn to this because the head element in this layout template was missing the `prefix="dcterms: http://purl.org/dc/terms/#"` attribute, so the `<meta name="dcterms.modified" ...>` element was invalid.  But it was also missing all the `apple-touch-icon`s and such.

The only problem with this is that the base layout includes the script snippet that reads the preferred theme from localStorage, but that's better dealt with another way.