Add 17 newly accessioned 88-note rolls
Note that nearly all of the 88-note rolls (not just the new ones) have modified .json files because of pending changes to the pianolatron-data backend processing so that the 'bad' (unusually sized) holes that are included for playback in the MIDI file also are highighted/hoverable in the viewer. And a smaller number of the older 88-note rolls have (beneficial) changes to their MIDI files because of upcoming improvements to the `roll-image-parser` code that are necessary to parse the new roll images properly.