Quickfix for 88-note roll zk199jz1564
Previously, this roll was one of the gen1 88-note rolls (actually it was the only one) with a strong-armed shift in tracker alignment/MIDI assignment hard-coded into the roll-wrangler script to make the sustain pedal holes be interpreted properly. Due to the recent improvements to the roll image parser code, this is no longer necessary!