Ape Spotlight header a bit more closely
The previous layout had very minor differences in the height of the navbar strip (the one with HOME - ROLL ARCHIVE FEATURES, etc.) compared to the SUPRA/Spotlight header. Of course the two sites' headers will differ in content, but it's probably desirable to avoid any noticeable shifts when switching between them in different browser tabs, even if you have to be paying close attention to register them.